ai can pop ulti
I am messing with the amount given for units dying. trying 2 rn, but in future a stat probably.

there is the potential for if you are winning to just keep winning if they push

also I was commanding units out of the range of ulti and playing cautiously some, though at one point I did look to see their progress