teaching player first itteration
I think the idea of making it look like it is drawn in mud is cool.

but I also though your unit buttons telling you the controls could be a fun idea too, for the other stuff, more context oriented

and probably don't auto command the units to the enemy base by default, a check on the bottom right?